+++ transforms/comp_kv_setter.py
@@ -3,10 +3,13 @@
 This is useful to compute spawnflags, or to adjust keyvalues when the target
 entity's options can't be set to a fixup variable.
 """
-from srctools import Angle, conv_int, conv_bool, Vec
+import re
+
+from srctools import Angle, conv_int, conv_bool, Vec, FGD
+from srctools.fgd import EntityDef
 from srctools.logger import get_logger
 
-from hammeraddons.bsp_transform import trans, Context
+from hammeraddons.bsp_transform import trans, Context, ent_description
 from hammeraddons.bsp_transform.common import check_control_enabled, get_multimode_value
 
 LOGGER = get_logger(__name__)
@@ -19,6 +22,7 @@ def kv_setter(ctx: Context) -> None:
         setter.remove()
         if not check_control_enabled(setter):
             continue
+        desc = ent_description(setter)
 
         mode = setter['mode', 'kv']
         if mode.casefold() == 'flags':
@@ -27,9 +31,9 @@ def kv_setter(ctx: Context) -> None:
             is_flags = False
         else:
             LOGGER.warning(
-                'Unknown mode ({}) for comp_kv_setter at ({})!',
+                'Unknown mode ({}) for {}!',
                 mode,
-                setter['origin'],
+                desc,
             )
             continue
 
@@ -47,39 +51,59 @@ def kv_setter(ctx: Context) -> None:
         elif conv_bool(setter['conv_ang']):
             kv_value = str(Vec.from_str(kv_value).to_angle())
 
+        conf_flag_mask: str | int
         if is_flags:
             try:
                 # Convert using Python's literal rules,
                 # to allow binary/hex literals.
-                flag_mask = int(kv_name, base=0)
+                conf_flag_mask = int(kv_name, base=0)
             except (TypeError, ValueError):
-                LOGGER.warning(
-                    'Invalid spawnflags mask for comp_kv_setter at ({})!\n'
-                    'Provide an decimal integer, 0x45 hex or 0b01101 binary value.',
-                    mode,
-                    setter['origin'],
-                )
-                continue
+                conf_flag_mask = normalise_flag(kv_name)
             flag_enabled = conv_bool(kv_value)
         else:
-            flag_mask = 1
+            conf_flag_mask = 1
             flag_enabled = False
             kv_name = kv_name.strip()
 
         if not is_flags and not kv_name and not setter.outputs:
             # We have nothing to do?
             LOGGER.warning(
-                'comp_kv_setter at ({}) is set to do nothing at all. '
+                '{} is set to do nothing at all. '
                 'Provide a keyvalue to set, spawnflag to change or '
                 'outputs to append.',
-                setter['origin'],
+                desc,
             )
 
         found_ent = None
+        flag_mask: int | None
+        fgd_flags: dict[str, int | None] = {}
 
         for found_ent in ctx.vmf.search(setter['target']):
             if is_flags:
                 spawnflags = conv_int(found_ent['spawnflags', '0'])
+                if isinstance(conf_flag_mask, str):
+                    # Try and lookup the FGD.
+                    found_ent_cls = found_ent['classname']
+                    try:
+                        flag_mask = fgd_flags[found_ent_cls]
+                    except KeyError:
+                        flag_mask = fgd_flags[found_ent_cls] = get_fgd_mask(found_ent_cls, conf_flag_mask)
+                        if flag_mask is None:
+                            LOGGER.warning(
+                                'Invalid spawnflags mask for {}!\n'
+                                'Provide the name of a flag, a decimal integer, 0x45 hex or 0b01101 binary value.',
+                                desc,
+                            )
+                            continue
+                        LOGGER.info(
+                            'FGD mask lookup: {0!r} in {1} = {2:#b} ({2})',
+                            conf_flag_mask, found_ent_cls, flag_mask,
+                        )
+                    else:
+                        if flag_mask is None:
+                            continue  # Don't repeat an error.
+                else:
+                    flag_mask = conf_flag_mask
                 if flag_enabled:
                     found_ent['spawnflags'] = spawnflags | flag_mask
                 else:
@@ -92,7 +116,28 @@ def kv_setter(ctx: Context) -> None:
 
         if found_ent is None:
             LOGGER.warning(
-                'No entities found named "{}" for comp_kv_setter at ({})!',
-                setter['target'],
-                setter['origin'],
+                'No entities found named "{}" for {}', setter['target'], desc,
             )
+
+
+def normalise_flag(flag: str) -> str:
+    """Normalise the flag, so punctuation doesn't affect results."""
+    flag = re.sub(r'[()]', '', flag.casefold())
+    return re.sub(r'[ \t]+', ' ', flag)
+
+
+def get_fgd_mask(ent_cls: str, search: str) -> int | None:
+    try:
+        ent_def = EntityDef.engine_def(ent_cls)
+    except KeyError:
+        LOGGER.warning('No FGD definition for entity {}!', ent_cls)
+        return None
+    try:
+        flags_kv = ent_def.kv['spawnflags']
+    except KeyError:
+        LOGGER.warning('No spawnflags in FGD definition for entity {}!', ent_cls)
+        return None
+    for (mask, name, default, flags) in flags_kv.flags_list:
+        if search == normalise_flag(name):
+            return mask
+    return None